auth_project/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
cryptography>=3.4.0
redis>=5.0.0
python-dotenv>=1.0.1
httpx>=0.27.0
aiohttp>=3.9.0
orjson>=3.10.0
pydantic[email]>=2.10.0